gjd-service/requirements.txt
fastapi
passlib
PyJWT
redis
uvicorn
sqlalchemy
python-dotenv